image_classifier/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
six
opencv-python
torch
torchvision
streamlit
pillow
torch
torchvision
kaggle
